JS: move TokenType to new file; optimize Bytes() for identifiers
