+++ js/lex.go
@@ -2,7 +2,6 @@
 package js
 
 import (
-	"strconv"
 	"unicode"
 	"unicode/utf8"
 
@@ -22,437 +21,6 @@ func IsIdentifierStart(b []byte) bool {
 func IsIdentifierEnd(b []byte) bool {
 	r, _ := utf8.DecodeLastRune(b)
 	return r == '$' || r == '\\' || r == '\u200C' || r == '\u200D' || unicode.IsOneOf(identifierContinue, r)
-}
-
-////////////////////////////////////////////////////////////////
-
-// TokenType determines the type of token, eg. a number or a semicolon.
-type TokenType uint16 // from LSB to MSB: 8 bits for tokens per category, 1 bit for numeric, 1 bit for punctuator, 1 bit for operator, 1 bit for identifier, 4 bits unused
-
-// TokenType values.
-const (
-	ErrorToken TokenType = iota // extra token when errors occur
-	WhitespaceToken
-	LineTerminatorToken // \r \n \r\n
-	CommentToken
-	CommentLineTerminatorToken
-	StringToken
-	TemplateToken
-	TemplateStartToken
-	TemplateMiddleToken
-	TemplateEndToken
-	RegExpToken
-)
-
-const (
-	NumericToken TokenType = 0x0100 + iota
-	DecimalToken
-	BinaryToken
-	OctalToken
-	HexadecimalToken
-	BigIntToken
-)
-
-const (
-	PunctuatorToken   TokenType = 0x0200 + iota
-	OpenBraceToken              // {
-	CloseBraceToken             // }
-	OpenParenToken              // (
-	CloseParenToken             // )
-	OpenBracketToken            // [
-	CloseBracketToken           // ]
-	DotToken                    // .
-	SemicolonToken              // ;
-	CommaToken                  // ,
-	QuestionToken               // ?
-	ColonToken                  // :
-	ArrowToken                  // =>
-	EllipsisToken               // ...
-)
-
-const (
-	OperatorToken TokenType = 0x0600 + iota
-	EqToken                 // =
-	EqEqToken               // ==
-	EqEqEqToken             // ===
-	NotToken                // !
-	NotEqToken              // !=
-	NotEqEqToken            // !==
-	LtToken                 // <
-	LtEqToken               // <=
-	LtLtToken               // <<
-	LtLtEqToken             // <<=
-	GtToken                 // >
-	GtEqToken               // >=
-	GtGtToken               // >>
-	GtGtEqToken             // >>=
-	GtGtGtToken             // >>>
-	GtGtGtEqToken           // >>>=
-	AddToken                // +
-	AddEqToken              // +=
-	IncrToken               // ++
-	SubToken                // -
-	SubEqToken              // -=
-	DecrToken               // --
-	MulToken                // *
-	MulEqToken              // *=
-	ExpToken                // **
-	ExpEqToken              // **=
-	DivToken                // /
-	DivEqToken              // /=
-	ModToken                // %
-	ModEqToken              // %=
-	BitAndToken             // &
-	BitOrToken              // |
-	BitXorToken             // ^
-	BitNotToken             // ~
-	BitAndEqToken           // &=
-	BitOrEqToken            // |=
-	BitXorEqToken           // ^=
-	AndToken                // &&
-	OrToken                 // ||
-	NullishToken            // ??
-	OptChainToken           // ?.
-
-	// unused in lexer
-	PosToken      // +a
-	NegToken      // -a
-	PreIncrToken  // ++a
-	PreDecrToken  // --a
-	PostIncrToken // a++
-	PostDecrToken // a--
-)
-
-const (
-	ReservedToken TokenType = 0x0800 + iota
-	AwaitToken
-	BreakToken
-	CaseToken
-	CatchToken
-	ClassToken
-	ConstToken
-	ContinueToken
-	DebuggerToken
-	DefaultToken
-	DeleteToken
-	DoToken
-	ElseToken
-	EnumToken
-	ExportToken
-	ExtendsToken
-	FalseToken
-	FinallyToken
-	ForToken
-	FunctionToken
-	IfToken
-	ImportToken
-	InToken
-	InstanceofToken
-	NewToken
-	NullToken
-	ReturnToken
-	SuperToken
-	SwitchToken
-	ThisToken
-	ThrowToken
-	TrueToken
-	TryToken
-	TypeofToken
-	VarToken
-	VoidToken
-	WhileToken
-	WithToken
-	YieldToken
-)
-
-const (
-	IdentifierToken TokenType = 0x1000 + iota
-	AsyncToken
-	ImplementsToken
-	InterfaceToken
-	LetToken
-	PackageToken
-	PrivateToken
-	ProtectedToken
-	PublicToken
-	StaticToken
-	OfToken
-	GetToken
-	SetToken
-	TargetToken
-	MetaToken
-	AsToken
-	FromToken
-)
-
-func IsNumeric(tt TokenType) bool {
-	return tt&0x0100 != 0
-}
-
-func IsPunctuator(tt TokenType) bool {
-	return tt&0x0200 != 0
-}
-
-func IsOperator(tt TokenType) bool {
-	return tt&0x0400 != 0
-}
-
-// IsIdentifierName matches IdentifierName, i.e. any identifier
-func IsIdentifierName(tt TokenType) bool {
-	return tt&0x1800 != 0
-}
-
-// IsReservedWord matches ReservedWord
-func IsReservedWord(tt TokenType) bool {
-	return tt&0x0800 != 0
-}
-
-// IsIdentifier matches Identifier, i.e. IdentifierName but not ReservedWord. Does not match yield or await.
-func IsIdentifier(tt TokenType) bool {
-	return tt&0x1000 != 0
-}
-
-// String returns the string representation of a TokenType.
-func (tt TokenType) String() string {
-	s := tt.Bytes()
-	if s == nil {
-		return "Invalid(" + strconv.Itoa(int(tt)) + ")"
-	}
-	return string(s)
-}
-
-var opBytes = [][]byte{
-	[]byte("Operator"),
-	[]byte("="),
-	[]byte("=="),
-	[]byte("==="),
-	[]byte("!"),
-	[]byte("!="),
-	[]byte("!=="),
-	[]byte("<"),
-	[]byte("<="),
-	[]byte("<<"),
-	[]byte("<<="),
-	[]byte(">"),
-	[]byte(">="),
-	[]byte(">>"),
-	[]byte(">>="),
-	[]byte(">>>"),
-	[]byte(">>>="),
-	[]byte("+"),
-	[]byte("+="),
-	[]byte("++"),
-	[]byte("-"),
-	[]byte("-="),
-	[]byte("--"),
-	[]byte("*"),
-	[]byte("*="),
-	[]byte("**"),
-	[]byte("**="),
-	[]byte("/"),
-	[]byte("/="),
-	[]byte("%"),
-	[]byte("%="),
-	[]byte("&"),
-	[]byte("|"),
-	[]byte("^"),
-	[]byte("~"),
-	[]byte("&="),
-	[]byte("|="),
-	[]byte("^="),
-	[]byte("&&"),
-	[]byte("||"),
-	[]byte("??"),
-	[]byte("?."),
-	[]byte("+"),
-	[]byte("-"),
-	[]byte("++"),
-	[]byte("--"),
-	[]byte("++"),
-	[]byte("--"),
-}
-
-// Bytes returns the string representation of a TokenType.
-func (tt TokenType) Bytes() []byte {
-	if IsOperator(tt) && tt <= PostDecrToken {
-		return opBytes[tt-OperatorToken]
-	}
-
-	switch tt {
-	case ErrorToken:
-		return []byte("Error")
-	case WhitespaceToken:
-		return []byte("Whitespace")
-	case LineTerminatorToken:
-		return []byte("LineTerminator")
-	case CommentToken:
-		return []byte("Comment")
-	case CommentLineTerminatorToken:
-		return []byte("CommentLineTerminator")
-	case StringToken:
-		return []byte("String")
-	case TemplateToken:
-		return []byte("Template")
-	case TemplateStartToken:
-		return []byte("TemplateStart")
-	case TemplateMiddleToken:
-		return []byte("TemplateMiddle")
-	case TemplateEndToken:
-		return []byte("TemplateEnd")
-	case RegExpToken:
-		return []byte("RegExp")
-	case NumericToken:
-		return []byte("Numeric")
-	case DecimalToken:
-		return []byte("Decimal")
-	case BinaryToken:
-		return []byte("Binary")
-	case OctalToken:
-		return []byte("Octal")
-	case HexadecimalToken:
-		return []byte("Hexadecimal")
-	case BigIntToken:
-		return []byte("BigInt")
-	case PunctuatorToken:
-		return []byte("Punctuator")
-	case OpenBraceToken:
-		return []byte("{")
-	case CloseBraceToken:
-		return []byte("}")
-	case OpenParenToken:
-		return []byte("(")
-	case CloseParenToken:
-		return []byte(")")
-	case OpenBracketToken:
-		return []byte("[")
-	case CloseBracketToken:
-		return []byte("]")
-	case DotToken:
-		return []byte(".")
-	case SemicolonToken:
-		return []byte(";")
-	case CommaToken:
-		return []byte(",")
-	case QuestionToken:
-		return []byte("?")
-	case ColonToken:
-		return []byte(":")
-	case ArrowToken:
-		return []byte("=>")
-	case EllipsisToken:
-		return []byte("...")
-	case IdentifierToken:
-		return []byte("Identifier")
-	case AwaitToken:
-		return []byte("await")
-	case BreakToken:
-		return []byte("break")
-	case CaseToken:
-		return []byte("case")
-	case CatchToken:
-		return []byte("catch")
-	case ClassToken:
-		return []byte("class")
-	case ConstToken:
-		return []byte("const")
-	case ContinueToken:
-		return []byte("continue")
-	case DebuggerToken:
-		return []byte("debugger")
-	case DefaultToken:
-		return []byte("default")
-	case DeleteToken:
-		return []byte("delete")
-	case DoToken:
-		return []byte("do")
-	case ElseToken:
-		return []byte("else")
-	case EnumToken:
-		return []byte("enum")
-	case ExportToken:
-		return []byte("export")
-	case ExtendsToken:
-		return []byte("extends")
-	case FalseToken:
-		return []byte("false")
-	case FinallyToken:
-		return []byte("finally")
-	case ForToken:
-		return []byte("for")
-	case FunctionToken:
-		return []byte("function")
-	case IfToken:
-		return []byte("if")
-	case ImportToken:
-		return []byte("import")
-	case InToken:
-		return []byte("in")
-	case InstanceofToken:
-		return []byte("instanceof")
-	case NewToken:
-		return []byte("new")
-	case NullToken:
-		return []byte("null")
-	case ReturnToken:
-		return []byte("return")
-	case SuperToken:
-		return []byte("super")
-	case SwitchToken:
-		return []byte("switch")
-	case ThisToken:
-		return []byte("this")
-	case ThrowToken:
-		return []byte("throw")
-	case TrueToken:
-		return []byte("true")
-	case TryToken:
-		return []byte("try")
-	case TypeofToken:
-		return []byte("typeof")
-	case VarToken:
-		return []byte("var")
-	case VoidToken:
-		return []byte("void")
-	case WhileToken:
-		return []byte("while")
-	case WithToken:
-		return []byte("with")
-	case YieldToken:
-		return []byte("yield")
-	case LetToken:
-		return []byte("let")
-	case StaticToken:
-		return []byte("static")
-	case ImplementsToken:
-		return []byte("implements")
-	case InterfaceToken:
-		return []byte("interface")
-	case PackageToken:
-		return []byte("package")
-	case PrivateToken:
-		return []byte("private")
-	case ProtectedToken:
-		return []byte("protected")
-	case PublicToken:
-		return []byte("public")
-	case AsToken:
-		return []byte("as")
-	case AsyncToken:
-		return []byte("async")
-	case FromToken:
-		return []byte("from")
-	case GetToken:
-		return []byte("get")
-	case MetaToken:
-		return []byte("meta")
-	case OfToken:
-		return []byte("of")
-	case SetToken:
-		return []byte("set")
-	case TargetToken:
-		return []byte("target")
-	}
-	return nil
 }
 
 ////////////////////////////////////////////////////////////////
